requirements.txt
# CET Admission System Requirements
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Image handling
Pillow==10.0.1

# Configuration
python-decouple==3.8

# Forms and UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

# Document generation
reportlab==4.0.7
openpyxl==3.1.2